src/test/java/com/nitesh/spring_test_project/service/EmployeeServiceTest.java
package com.nitesh.spring_test_project.service;

import com.nitesh.spring_test_project.repository.EmployeeRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import static org.assertj.core.api.Assertions.assertThat;

@ExtendWith({SpringExtension.class})
@ExtendWith(MockitoExtension.class)
class EmployeeServiceTest {

    @Mock
    private EmployeeRepository employeeRepository;

    @InjectMocks
    private EmployeeService employeeService;

//    @BeforeEach
//    public void setup() {
//        employeeRepository = Mockito.mock(EmployeeRepository.class);
//        employeeService = new EmployeeService(employeeRepository);
//    }


    @Test
    void testBeansAreNotNull() {
        assertThat(employeeRepository).isNotNull();
        assertThat(employeeService).isNotNull();
        assertThat(employeeService.getEmployeeRepository()).isNotNull();

    }

}